remove separate createOpenTagMarkup file for now since everything should be copy pasted into a single file
